pyproject.toml
[project]
name = "detect-yolo-cpu-v2"
version = "2.0.0"
description = "Détection d'objets YOLO CPU avec zones et intégration Home Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.8.0",
    "shapely>=2.0.0",
    "watchdog>=3.0.0",
    "paho-mqtt>=1.6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "structlog>=23.2.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
filterwarnings = [
    "ignore::UserWarning:torch.cuda"  # ← AJOUTER cette ligne
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
]
